Create several new types
