+++ src/types/Node.ts
@@ -1,0 +1,7 @@
+type Node = {
+  id: string;
+  name: string;
+  namespace: string;
+}
+
+export default Node;
