+++ src/types/RosStdMsgsString.ts
@@ -1,0 +1,5 @@
+type RosStdMsgString = {
+  data: string;
+};
+
+export default RosStdMsgString;
