--- src/types/StdMsgString.ts
@@ -1,3 +1,0 @@
-export type StdMsgString = {
-  data: string;
-};
