+++ src/types/Topic.ts
@@ -1,0 +1,9 @@
+type Topic = {
+  id: string;
+  name: string;
+  namespace: string;
+  sourceId: string;
+  destId: string;
+};
+
+export default Topic;
